size param to perfstat functions is size_t not int

+++ src/os/aix/perfstat/sigar_perfstat.c
@@ -6,14 +6,14 @@
  *  desired_number Must be set to 1."
  * so we just hardcode that in our wrapper.
  */
-int sigar_perfstat_cpu_total(perfstat_cpu_total_t *cpu_total, int size)
+int sigar_perfstat_cpu_total(perfstat_cpu_total_t *cpu_total, size_t size)
 {
     return perfstat_cpu_total(NULL, cpu_total, size, 1);
 }
 
 int sigar_perfstat_cpu(perfstat_id_t *id,
                        perfstat_cpu_t *cpu,
-                       int size, int num)
+                       size_t size, int num)
 {
     return perfstat_cpu(id, cpu, size, num);
 }
